binance-api/requirements.txt
# lightweightcharts
python-binance
websocket_client
config
TA-Lib
numpy
backtrader
vectorbt
flask
flask-cors
backtrader
pandas

